@@ -3,7 +3,7 @@ import Foundation
 /// Represents a licensed component.
 public struct LicensedComponent: Hashable, Identifiable {
     @usableFromInline
-    struct _CopyrightDetails: Hashable, Sendable {
+    struct _CopyrightDetails: Sendable, Hashable {
         @usableFromInline
         let licensedProduct: String
         @usableFromInline
@@ -104,9 +104,9 @@ extension LicensedComponent {
     /// For some "known" licenses, this package already bundles the text (see ``LicensedComponent/License``).
     /// Each license can have a short text and the full license text. The short text should not exceed a certain length.
     /// If the full license text fits into the short text, ``LicenseTexts/full`` can be set to `nil`.
-    public struct LicenseTexts: Hashable, Sendable {
+    public struct LicenseTexts: Sendable, Hashable {
         /// The resolved (enriched) version of the license texts where placeholders have been replaced with the corresponding values.
-        public struct Resolved: Hashable, Sendable {
+        public struct Resolved: Sendable, Hashable {
             /// The resolved short license text.
             public let short: String
             /// The resolved full license text.
@@ -123,7 +123,7 @@ extension LicensedComponent {
         ///
         /// - Note: If you provide a custom text and don't have any placeholders,
         ///         pass `false` to `hasPlaceholders` to prevent any string manipulation.
-        public struct Text: Hashable, Sendable {
+        public struct Text: Sendable, Hashable {
             /// The string of this text.
             public let string: String
             /// Whether `string` contains placeholders.
@@ -170,21 +170,21 @@ extension LicensedComponent {
 
 extension LicensedComponent {
     /// Describes a license.
-    public enum License: Hashable, Sendable {
+    public enum License: Sendable, Hashable {
         /// The Apache license versions.
-        public enum Apache: Hashable, Sendable {
+        public enum Apache: Sendable, Hashable {
             /// Version 2.0 of the Apache license.
             case v2
         }
 
         /// The BSD license versions.
-        public enum BSD: Hashable, Sendable {
+        public enum BSD: Sendable, Hashable {
             /// The 3-Clause version of the BSD license.
             case threeClause
         }
 
         /// The GPL license versions.
-        public enum GPL: Hashable, Sendable {
+        public enum GPL: Sendable, Hashable {
             /// Version 3.0 of the GPL license.
             case v3
         }
